Rename flow export app/program to ipfix
This is shorter and it's more obvious what it's about. Also if
in the future we make a flow collector app, we can put it in the
same ipfix folder.
