+++ src/program/ipfix/README
@@ -1,0 +1,6 @@
+Usage:
+  snabb ipfix probe
+
+Use --help for per-command usage.
+Example:
+    snabb ipfix probe --help
